backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
web3==6.15.1
ipfshttpclient==0.7.0
python-jose==3.3.0
requests==2.31.0
pytest==8.0.2
black==24.1.1
flake8==7.0.0
python-multipart==0.0.9
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
python-magic==0.4.27 